Formatted all the specs
