--- Tests/PushSpec.swift
@@ -1,18 +1,0 @@
-//
-//  PushSpec.swift
-//  SwiftPhoenixClient
-//
-//  Created by Daniel Rees on 5/18/18.
-//
-
-import Quick
-import Nimble
-@testable import SwiftPhoenixClient
-
-class PushSpec: QuickSpec {
-    
-    override func spec() {
-        
-    }
-}
-
